Split writing of IP4Datagram header attributes into NetworkByteBlock into single steps to have all attributes in the right byte order finally, changed size of IP4DataPart to correct data type.

+++ src/kernel/network/internet/IP4DataPart.h
@@ -36,7 +36,7 @@ public:
 
     virtual uint8_t copyDataTo(NetworkByteBlock *byteBlock) = 0;
 
-    virtual uint16_t getLengthInBytes() = 0;
+    virtual size_t getLengthInBytes() = 0;
 
     virtual IP4ProtocolType getIP4ProtocolType() = 0;
 };
